backend/requirements.txt
fastapi
uvicorn
motor
python-multipart
httpx
cloudinary
google-generativeai
python-dotenv
pydantic[email]
requests
email-validator
